requirements.txt
beautifulsoup4
bitarray
cryptography
flask
matplotlib
numpy
pandas
pdoc3
psutil
pymysql
python-dotenv
requests
simphile
sklearn
numpy
pdoc3
matplotlib
flask
cryptography
pyjwt
bcrypt
flask_cors
geoip2
suffix-tree
